LayoutTests/fast/dom/shadow/shadowdom-for-image-event-expected.txt

Tests to ensure that an img element receives a load event, but an inner image element in shadow DOM subtree does not receive.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.currentTarget.id is "host"
